add conccurent scaler/scheduler logic

+++ simulator/data_structs.go
@@ -8,7 +8,7 @@ type functionProfile struct {
 	Trigger     string
 	PerMinute   [1440]int
 	AvgDuration int
-	AvgMemory   int
+	AvgMemory   int64
 }
 
 type functionExecutionDuration struct {
@@ -32,13 +32,13 @@ type appMemory struct {
 	owner                string
 	app                  string
 	count                int
-	average              int
-	percentileAverage1   int
-	percentileAverage5   int
-	percentileAverage25  int
-	percentileAverage50  int
-	percentileAverage75  int
-	percentileAverage95  int
-	percentileAverage99  int
-	percentileAverage100 int
+	average              int64
+	percentileAverage1   int64
+	percentileAverage5   int64
+	percentileAverage25  int64
+	percentileAverage50  int64
+	percentileAverage75  int64
+	percentileAverage95  int64
+	percentileAverage99  int64
+	percentileAverage100 int64
 }
